LayoutTests/platform/chromium-mac/platform/mac/editing/input/firstrectforcharacterrange-plain-expected.txt

EDITING DELEGATE: shouldBeginEditingInDOMRange:range from 0 of BODY > HTML > #document to 2 of BODY > HTML > #document
EDITING DELEGATE: webViewDidBeginEditing:WebViewDidBeginEditingNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
123456
10,10,0,18
18,10,0,18
58,10,0,18
58,10,0,18
10,10,8,18
50,10,8,18
50,10,8,18
50,10,8,18
10,10,48,18
10,10,48,18
10,10,0,18
18,10,0,18
null
null
null
null
